Versión QA 2.0
SE agrego validación de rut, se cambio el numero de intentos a 2 y se agrego un nuevo valor con true o false en la autenticación

+++ routes/bot.js
@@ -46,7 +46,8 @@ router.post('/message', async (req, res) => {
     "messages": [],
     "additionalInfo": {
       "key":"RUT",
-      "RUT":"1-9"
+      "RUT":"1-9",
+      "authValidity": ""
     }
   }
   
@@ -65,6 +66,19 @@ router.post('/message', async (req, res) => {
   var msj_aut_erroneo = await controlador.funciones.cargar_aut_errroneo();
 
   var msj_preguntas_rut = await controlador.funciones.cargar_preguntas_rut();
+
+  var local_function = {
+    no_autenticado : async function()
+    {
+      console.log("[local_function] :: [no_autenticado]");
+
+      resultado.action = msj_aut_exitosa.action;                    
+      resultado.messages.push(msj_aut_exitosa.messages[1]);
+      resultado.additionalInfo.authValidity = false;
+
+      localStorage.clear();     
+    }
+  }
 
   if(apiVersion !== '' && typeof apiVersion !== "undefined") 
   {
@@ -86,34 +100,37 @@ router.post('/message', async (req, res) => {
                 {
                   var valida_vigencia = await controlador.funciones.valida_vigencia(user.id);
 
-                  if(valida_vigencia === true)
+                  localStorage.setItem("valida_vigencia"+conversationID, valida_vigencia.id);
+
+                  console.log("NO ES CLIENTE **************", valida_vigencia.authValidity, typeof valida_vigencia.authValidity);
+
+                  if(valida_vigencia.authValidity == true)
                   {
+                    console.log("[valida_vigencia] :: [TRUE]");
+
                     resultado.action = msj_aut_exitosa.action;
-                    resultado.timeoutInactivity = 9999;
-
-                    for (var i = 0; i < msj_aut_exitosa.messages.length; i++)
-                    {
-                      resultado.messages.push(msj_aut_exitosa.messages[i]);
-                    }
-
+                    resultado.action["type"] = "continue"; 
+                    resultado.messages.push(msj_aut_exitosa.messages[0]);   
                     resultado.messages[0].text = resultado.messages[0].text.replace("<nombre cliente>", user.name);
 
-                    localStorage.clear();
+                    localStorage.setItem("pregunta_rut"+conversationID, ["transferir", false, valida_vigencia.id]);
+                    localStorage.setItem("bot_bci_"+conversationID, "cliente");
                   }
-                  else if(valida_vigencia === 99)
+                  else if(valida_vigencia.authValidity === 99)
                   {
+
                     resultado.action = msj_no_cliente.action;
                     resultado.messages.push(msj_no_cliente.messages[0]);
 
                     localStorage.clear();
                   }  
-                  else
+                  else if(valida_vigencia.authValidity == false)
                   {
                     resultado.action = msj_inicial.action;
                     resultado.messages.push(msj_inicial.messages[0]);
 
                     localStorage.setItem("bot_bci_"+conversationID, "cliente");
-                    localStorage.setItem("pregunta_rut"+conversationID, ["rut",false]);
+                    localStorage.setItem("pregunta_rut"+conversationID, ["rut",false]);                    
                   }                  
                 }
 
@@ -124,8 +141,7 @@ router.post('/message', async (req, res) => {
                   console.log("pregunta_rut :: " + pregunta_rut);
 
                   if(pregunta_rut[0] == "rut" && pregunta_rut[1] == "false" )
-                  {
-                    
+                  {                    
                     resultado.action = msj_preguntas_rut.action;
                     resultado.action.saveHistory = false;
                     resultado.messages.push(msj_preguntas_rut.messages[0]);
@@ -136,35 +152,44 @@ router.post('/message', async (req, res) => {
                   else if(pregunta_rut[0] == "rut" && pregunta_rut[1] == "true" )
                   {
                     var bandera_vali = await controlador.funciones.validación_campos_rut(pregunta_rut[0], mensaje);
+                    var rut_vigencia =  localStorage.getItem("valida_vigencia"+conversationID);
+
+                    console.log("[Cliente ingreso RUT] :: ", rut_vigencia.replace(/-/g,""), mensaje);
 
                     if(bandera_vali)
                     {
-                      var  respuesta_rut = localStorage.getItem("respuesta_rut"+conversationID);
-
-                      if(respuesta_rut == null)
+                      if(/*rut_vigencia.replace(/-/g,"") == mensaje*/ true)
                       {
-                        localStorage.setItem("respuesta_rut"+conversationID, mensaje); 
+                        var  respuesta_rut = localStorage.getItem("respuesta_rut"+conversationID);
+
+                        if(respuesta_rut == null)
+                        {
+                          localStorage.setItem("respuesta_rut"+conversationID, mensaje); 
+                        }
+
+                        resultado.action = msj_preguntas_rut.action;
+                        resultado.action.saveHistory = false;
+                        resultado.messages.push(msj_preguntas_rut.messages[1]);
+
+                        localStorage.setItem("pregunta_rut"+conversationID, ["numSerie",true]);
                       }
-
-                      resultado.action = msj_preguntas_rut.action;
-                      resultado.action.saveHistory = false;
-                      resultado.messages.push(msj_preguntas_rut.messages[1]);
-
-                      localStorage.setItem("pregunta_rut"+conversationID, ["numSerie",true]);
+                      else // Cliente no autenticado
+                      {
+                        await local_function.no_autenticado();
+                      }                      
                     }
                     else
                     {
                       resultado.action = msj_preguntas_rut.action;
                       resultado.action.saveHistory = false;
                       resultado.messages.push(msj_preguntas_rut.messages[0]);
-                    }
-                      
+                    }                      
                   }
 
                   if(pregunta_rut[0] == "numSerie" && pregunta_rut[1] == "true" )
                   {
                     var bandera_vali = await controlador.funciones.validación_campos_rut(pregunta_rut[0], mensaje);
-
+                    
                     if(bandera_vali)
                     {
                       var respuesta_rut = localStorage.getItem("respuesta_rut"+conversationID);
@@ -174,20 +199,18 @@ router.post('/message', async (req, res) => {
                       if(axios_CEDU.status)
                       {
                         if(axios_CEDU.result == "10000")
-                        {
+                        {                         
                           resultado.action = msj_aut_exitosa.action;
-                          resultado.timeoutInactivity = 9999;
-
-                          for (var i = 0; i < msj_aut_exitosa.messages.length; i++)
-                          {
-                            resultado.messages.push(msj_aut_exitosa.messages[i]);
-                          }
-
+                          resultado.action["type"] = "continue";
+                          resultado.action.saveHistory = false;
+                          resultado.messages.push(msj_aut_exitosa.messages[0]);   
                           resultado.messages[0].text = resultado.messages[0].text.replace("<nombre cliente>", user.name);
 
-                          var update_vigencia = await controlador.funciones.update_vigencia(user.id, respuesta_rut);
+                          localStorage.setItem("pregunta_rut"+conversationID, ["transferir",true]);
 
-                          localStorage.clear();
+                          var rut_vigencia =  localStorage.getItem("valida_vigencia"+conversationID);
+
+                          var update_vigencia = await controlador.funciones.update_vigencia(user.id, rut_vigencia);
                         }
                         else
                         {
@@ -204,9 +227,8 @@ router.post('/message', async (req, res) => {
 
                           console.log("Resultado de CEDU :::::::: " + axios_CEDU.result);
 
-                          if((axios_CEDU.result == "10001" || axios_CEDU.result == "10005" || axios_CEDU.result == "10008") &&  parseInt(num_intentos) <= 2 )
-                          {
-                            
+                          if((axios_CEDU.result == "10001" || axios_CEDU.result == "10005" || axios_CEDU.result == "10008") &&  parseInt(num_intentos) <= 1 )
+                          {                            
                             resultado.action = msj_aut_erroneo.action;
 
                             for (var i = 0; i < msj_aut_erroneo.messages.length; i++)
@@ -217,27 +239,16 @@ router.post('/message', async (req, res) => {
                             localStorage.setItem("intento"+conversationID, num_intentos);
                             localStorage.setItem("pregunta_rut"+conversationID, ["intentar", true])
                           }
-                          else
+                          else // cambiar valor a  no autenticado en el contexto
                           {
-                            // cambiar valor a  no autenticado en el contexto
-                            
-                            resultado.action = msj_aut_exitosa.action;
-                            resultado.timeoutInactivity = 9999;                          
-                            resultado.messages.push(msj_aut_exitosa.messages[1]);
-                            localStorage.clear();
+                            await local_function.no_autenticado();
                           }
                         }
                       }
-                      else
+                      else // cambiar valor a  no autenticado en el contexto
                       {
-                        // cambiar valor a  no autenticado en el contexto
-                            
-                        resultado.action = msj_aut_exitosa.action;
-                        resultado.timeoutInactivity = 9999;
-                        resultado.messages.push(msj_aut_exitosa.messages[1]);
-
-                        localStorage.clear();
-                      }
+                        await local_function.no_autenticado();
+                      }                      
                     }
                     else
                     {
@@ -247,8 +258,23 @@ router.post('/message', async (req, res) => {
                     }                                     
                   }
 
+                  if(pregunta_rut[0] === "transferir" && pregunta_rut[2] !== "")
+                  {
+                    if(pregunta_rut[1] == "true") // Viene autenticado
+                    {
+                      resultado.action = msj_aut_exitosa.action;
+                      resultado.messages.push(msj_aut_exitosa.messages[1]);
+                      resultado.additionalInfo.authValidity = true;
+                      localStorage.clear();
+                    }
+                    else
+                    {
+                      localStorage.setItem("pregunta_rut"+conversationID, ["transferir", true, pregunta_rut[2]]);
+                    }                                         
+                  }
+
                   // Intentos
-                  if(pregunta_rut[0] === "intentar" && pregunta_rut[1] == "true" && (mensaje.toLowerCase() == "si" || mensaje.toLowerCase() == "ok"))
+                  if(pregunta_rut[0] === "intentar" && pregunta_rut[1] == "true" && (mensaje.toLowerCase() == "si" || mensaje.toLowerCase() == "1"))
                   {
                     resultado.action = msj_preguntas_rut.action;
                     resultado.action.saveHistory = false;
@@ -257,15 +283,10 @@ router.post('/message', async (req, res) => {
                     localStorage.setItem("pregunta_rut"+conversationID, ["rut",true]);  
                   }
                   
-                  if(pregunta_rut[0] === "intentar" && mensaje.toLowerCase() != "si" && mensaje.toLowerCase() != "ok")
+                  if(pregunta_rut[0] === "intentar" && mensaje.toLowerCase() != "si" && mensaje.toLowerCase() != "1")
                   {
-                    console.log("entro a intentar con un no :: " + mensaje +" :: "+  mensaje.toLowerCase() != "si" || mensaje.toLowerCase() != "ok");
-                    
-                    resultado.action = msj_aut_exitosa.action;
-                    resultado.timeoutInactivity = 9999;
-                    resultado.messages.push(msj_aut_exitosa.messages[1]);
-
-                    localStorage.clear();
+                    console.log("entro a intentar con un no :: " + mensaje +" :: "+  mensaje.toLowerCase() != "si" || mensaje.toLowerCase() != "1");
+                    await local_function.no_autenticado();
                   }                  
                 }
               }
